refactor: view 담당 클래스명 수정
1. view 패키지 생성해서 view 관련 클래스들 이동시킴
