+++ src/main/java/baseball/Application.java
@@ -1,14 +1,14 @@
 package baseball;
 
 import java.util.Scanner;
-import utils.Input;
+import view.InputView;
 
 public class Application {
 
     public static void main(String[] args) {
         final Scanner scanner = new Scanner(System.in);
 
-        Input.scanner = scanner;
+        InputView.scanner = scanner;
         NumberBaseBallGame game = new NumberBaseBallGame();
         game.run();
     }
